refactoring out common code, updating tailwind config with accordion shadcn stuff
