--- components/blocks/imageTextBlock/listItem-schema.tsx
@@ -1,22 +1,0 @@
-import { IconPickerInput } from "../sharedTinaFields/iconSelector";
-
-export const listItemSchema = [
-  {
-    type: "string",
-    label: "Heading",
-    name: "heading",
-  },
-  {
-    type: "string",
-    label: "Description",
-    name: "description",
-  },
-  {
-    type: "string",
-    label: "Icon",
-    name: "icon",
-    ui: {
-      component: IconPickerInput,
-    },
-  },
-];
