+++ components/blocks/index.ts
@@ -39,7 +39,8 @@ import { googleMapsSchema } from "./googleMapsWrapper";
 import { gridLayoutSchema } from "./gridLayout";
 import { heroBlockSchema } from "./hero";
 import { horizontalBlockSchema } from "./horizontalCard";
-import { ImageTextBlockSchema } from "./imageTextBlock/imageTextBlock.schema";
+import { AccordionSchema } from "./imageComponents/accordionBlock/accordionSchema";
+import { ImageTextBlockSchema } from "./imageComponents/imageTextBlock/imageTextBlock.schema";
 import { internalCarouselBlockSchema } from "./internalCarousel";
 import { jotFormEmbedSchema } from "./jotFormEmbed";
 import { newslettersTableBlockSchema } from "./newslettersTable";
@@ -109,4 +110,5 @@ export const pageBlocks: Template[] = [
   videoEmbedBlockSchema,
   inlineJotFormSchema,
   ImageTextBlockSchema,
+  AccordionSchema,
 ];
