+++ package.json
@@ -61,6 +61,7 @@
     "@microsoft/applicationinsights-web": "^3.3.4",
     "@next/bundle-analyzer": "^15.0.3",
     "@next/third-parties": "^14.2.15",
+    "@radix-ui/react-accordion": "^1.2.1",
     "@radix-ui/react-dropdown-menu": "^2.1.2",
     "@radix-ui/react-slot": "^1.1.0",
     "@tailwindcss/typography": "^0.5.15",
