+++ public/sitemap.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<sitemapindex xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
+<sitemap><loc>ssw.com.au/sitemap-0.xml</loc></sitemap>
+<sitemap><loc>https://www.ssw.com.au/people/sitemap-index.xml</loc></sitemap>
+<sitemap><loc>https://www.ssw.com.au/rules/sitemap-index.xml</loc></sitemap>
+<sitemap><loc>https://www.ssw.com.au/archive/sitemap-index.xml</loc></sitemap>
+</sitemapindex>
